refactor(renderer/api/weather): move interfaces to co-located types/ folder
