+++ src/renderer/api/weather/types/index.ts
@@ -1,0 +1,35 @@
+/*
+ * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
+ * https://github.com/JNTMTMTM/eIsland
+ *
+ * Copyright (C) 2026 JNTMTMTM
+ * Copyright (C) 2026 pyisland.com
+ *
+ * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+/**
+ * @file index.ts
+ * @description 天气模块类型统一导出
+ * @author 鸡哥
+ */
+
+export type { WeatherApiConfig } from './WeatherApiConfig';
+export type { WeatherAlertLocation, WeatherAlertSummary, StartupWeatherAlertPayload } from './WeatherAlert';
+export type { LocationInfo } from './LocationInfo';
+export type {
+  DistrictQueryParams,
+  DistrictItem,
+  DistrictResolvedLocation,
+  DistrictQueryResult,
+} from './District';
